add grades calc
