+++ frontend/src/components/ManageTeam/ManageTeam.tsx
@@ -20,6 +20,7 @@ export interface User {
   surname: string;
   status: string;
   averageGrade: number;
+  maxGrade: number;
 };
 
 export interface TeamInfo {
